added more optimizations

+++ client/src/components/ClearButton.jsx
@@ -1,19 +1,15 @@
 import { useSelector, useDispatch } from "react-redux"
 import { thunks as todosThunks } from "../global/slices/todosSlice"
-import { actions as todosActions } from "../global/slices/todosSlice"
 
 export default function ClearButton () {
     const { todos } = useSelector(s => s.todos)
     const dispatch = useDispatch()
 
     const clearTodos = () => {
-        if (window.confirm("Delete All Todos?")) {
-            dispatch(todosThunks.deleteAll())
-            dispatch(todosActions.clearTodos())
-        }
+        if (window.confirm("Delete All Todos?")) dispatch(todosThunks.deleteAll())
     }
 
-    if (todos.length === 0) return <></>//don't render if there are no todos
+    if (todos.length === 0) return <></>
     return (
         <button 
             onClick={clearTodos}
